+++ frontend/src/components/Chat/ChatWindow.tsx
@@ -10,8 +10,8 @@ import MessageList from './MessageList'
 import ChatInputArea from './ChatInputArea'
 import ConversationPanel from './ConversationPanel'
 import ConverterPanel from './ConverterPanel'
-import type { PieceConversion } from './ConverterPanel'
-import { PIECE_TYPE_TO_DATA_TYPE } from './ConverterPanel'
+import type { PieceConversion } from './converterTypes'
+import { PIECE_TYPE_TO_DATA_TYPE } from './converterTypes'
 import LabelsBar from '../Labels/LabelsBar'
 import type { ChatInputAreaHandle } from './ChatInputArea'
 import { attacksApi } from '../../services/api'
